requirements.txt
streamlit==1.38.0
requests==2.32.3
beautifulsoup4==4.12.3
scrapy==2.11.2
transformers==4.44.2
torch==2.4.1
sentence-transformers==3.1.1
faiss-cpu==1.8.0
bibtexparser==2.0.0